pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openjarvis"
version = "1.0.0"
description = "OpenJarvis — modular AI assistant backend with composable intelligence pillars"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8",
    "httpx>=0.27",
    "rich>=13",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5",
    "respx>=0.22",
    "ruff>=0.4",
]
inference-ollama = []
inference-vllm = []
inference-llamacpp = []
inference-cloud = [
    "openai>=1.30",
    "anthropic>=0.30",
]
inference-google = [
    "google-genai>=1.0",
]
tools-search = [
    "tavily-python>=0.3",
]
memory-faiss = [
    "faiss-cpu>=1.7",
    "sentence-transformers>=2.2",
    "numpy>=1.24",
]
memory-colbert = [
    "colbert-ai>=0.2",
    "torch>=2.0",
]
memory-pdf = ["pdfplumber>=0.10"]
memory-bm25 = ["rank-bm25>=0.2.2"]
server = [
    "fastapi>=0.110",
    "uvicorn>=0.30",
    "pydantic>=2.0",
]
agents = []
learning = []
openclaw = []

[project.scripts]
jarvis = "openjarvis.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/openjarvis"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "live: requires running inference engine",
    "cloud: requires cloud API keys",
    "nvidia: requires NVIDIA GPU",
    "amd: requires AMD GPU",
    "apple: requires Apple Silicon",
    "slow: long-running test",
]

[tool.ruff]
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
